support for dcc file size up to 4GB (before there was a limit of 2.4GB)
git-svn-id: https://svn.kvirc.de/svn/branches/kvirc/3.4@2619 17fca916-40b9-46aa-a4ea-0a15b648b75c

+++ src/modules/dcc/broker.h
@@ -94,7 +94,7 @@ public:
 	void sendFileManage(KviDccDescriptor * dcc);
 
 	bool handleResumeAccepted(const char * filename,const char * port,const char * szZeroPortTag);
-	bool handleResumeRequest(KviDccRequest * dcc,const char * filename,const char * port,unsigned int filePos,const char * szZeroPortTag);
+	bool handleResumeRequest(KviDccRequest * dcc,const char * filename,const char * port,unsigned long filePos,const char * szZeroPortTag);
 
 public slots:
 	void rsendExecute(KviDccBox * box,KviDccDescriptor * dcc);
